Rename signal.h to avoid namespace conflict on mingw
