+++ .github/workflows/build.yml
@@ -103,8 +103,8 @@ jobs:
           cmake \
             -S . \
             -B . \
-            -DCMAKE_BUILD_TYPE=${{ matrix.config.build_type }} \
             -G "${{ matrix.config.generators }}" \
+            -DBUILD_EXAMPLES=ON \
             -DCMAKE_INSTALL_PREFIX:PATH=instdir
 
       - name: Build
